+++ python/ray/data/_internal/execution/autoscaler/autoscaler.py
@@ -3,10 +3,10 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from ray._private.ray_constants import env_float
-from ray.data._internal.execution.interfaces.execution_options import ExecutionResources
 from ray.util.annotations import DeveloperAPI
 
 if TYPE_CHECKING:
+    from ray.data._internal.execution.interfaces.execution_options import ExecutionResources
     from ray.data._internal.execution.resource_manager import ResourceManager
     from ray.data._internal.execution.streaming_executor_state import Topology
 
@@ -82,6 +82,6 @@ class Autoscaler(ABC):
         ...
 
     @abstractmethod
-    def get_total_resources(self) -> ExecutionResources:
+    def get_total_resources(self) -> "ExecutionResources":
         """Get the total resources that are available to this data execution."""
         ...
